backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
Pillow==12.1.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
gunicorn==21.2.0
